* Misc cleanup based on Intellij code analysis.

+++ dslink-v2-api/src/main/java/org/iot/dsa/node/DSElement.java
@@ -136,7 +136,7 @@ public abstract class DSElement extends DSValue {
      * Creates an DSIObject representation of the primitive.
      */
     public static DSElement make(int arg) {
-        return DSLong.valueOf((long) arg);
+        return DSLong.valueOf(arg);
     }
 
     /**
