+++ dslink-v2-api/src/main/java/org/iot/dsa/node/DSInt.java
@@ -128,12 +128,12 @@ public class DSInt extends DSValue implements Comparable<DSINumber>, DSIMetadata
 
     @Override
     public int toInt() {
-        return (int) value;
+        return value;
     }
 
     @Override
     public long toLong() {
-        return (long) value;
+        return value;
     }
 
     @Override
